requirements.txt
asgiref==3.11.0
Django==5.2.8
sqlparse==0.5.4

djangorestframework